+++ components/@dopt/react/contextual-assistant/src/contextual-assistant/index.ts
@@ -1,0 +1,7 @@
+export * from './answer';
+export * from './content';
+export * from './header';
+export * from './highlights';
+export * from './popover';
+export * from './provider';
+export * from './title';