+++ components/@dopt/react/checklist/sandbox/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @dopt/react-checklist-sandbox
 
+## 0.0.32
+
+### Patch Changes
+
+- @dopt/react-checklist@9.0.1
+- @dopt/react-modal@9.0.1
+
 ## 0.0.31
 
 ### Patch Changes