+++ components/@dopt/semantic-data-layer/card/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @dopt/semantic-data-layer-card
 
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies [eaf9872d3]
+  - @dopt/block-api-types@1.3.2
+
 ## 1.2.0
 
 ### Minor Changes